+++ BiomeConqueror/BiomeConqueror.cs
@@ -15,7 +15,7 @@ namespace BiomeConqueror
     {
         public const string GUID = "Turbero.BiomeConqueror";
         public const string NAME = "Biome Conqueror";
-        public const string VERSION = "1.1.3";
+        public const string VERSION = "1.2.0";
         
         private readonly Harmony harmony = new Harmony(GUID);